fix: relocate capped watermark validation for module limits
